Malformed packets should not be able to crash the DataReader
If the packet length integer is negative, the JVM will throw an
exception. This exception is not catched and will crash the DataReader
thread silently. The PacketHandler class will still think that the
DataReader is running.

Now, the DataReader checks first whether the length is a valid - higher
than zero - number.